requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-filter==22.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
exceptiongroup==1.0.0
iniconfig==1.1.1
kombu==5.2.4
model-bakery==1.9.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
